fix: delete bug with right answer
